improve the scenario and the buidl.yml and simviewer
