+++ CMakeLists.txt
@@ -219,7 +219,7 @@ include(FetchContent)
 FetchContent_Declare(
         polyscope
         GIT_REPOSITORY https://github.com/nmwsharp/polyscope.git
-        GIT_TAG        v2.4.0
+        GIT_BRANCH master
         GIT_PROGRESS   YES
 )
