+++ imgui.ini
@@ -3,26 +3,26 @@ Pos=853,-16
 Size=856,679
 
 [Window][Simulation Controls]
-Pos=29,568
+Pos=55,314
 Size=440,304
 
 [Window][Performance]
-Pos=3,363
+Pos=707,463
 Size=246,250
 
 [Window][Debug]
-Pos=853,180
+Pos=75,23
 Size=158,102
 
 [Window][Scenarios]
-Pos=433,424
+Pos=932,110
 Size=324,558
 
 [Window][Scene Hierarchy]
 Pos=853,52
-Size=124,60
+Size=124,82
 
 [Window][Solver Settings]
-Pos=528,74
-Size=278,334
+Pos=507,147
+Size=278,510
